# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /policies/{policyType}:
    get:
      operationId: getPolicy
      tags: [policies]
      summary: Get one published policy document
      parameters:
        - name: policyType
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Published policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "404":
          description: Policy not found
...
  /policy-acceptances:
    post:
      operationId: createPolicyAcceptance
      tags: [policies]
      summary: Record that the signed-in user accepted a policy in a given context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyAcceptanceInput"
      responses:
        "201":
          description: Acceptance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyAcceptance"
        "401":
          description: Authentication required
...
  /admin/listings/{id}/moderation:
    patch:
      operationId: moderateListing
      tags: [moderation]
      summary: Set a listing's granular moderation workflow state
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingModerationInput"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminListing"
...
  /admin/policies/{policyType}:
    patch:
      operationId: updateAdminPolicy
      tags: [policies]
      summary: Update and publish a policy document (admin only)
      parameters:
        - name: policyType
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyUpdate"
      responses:
        "200":
          description: Updated policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPolicy"
  /admin/site-settings:
    patch:
      operationId: updateAdminSiteSettings
      tags: [policies]
      summary: Update business/contact configuration (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated site settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
...
components:
...
  schemas:
...
    ListingModerationInput:
      type: object
      properties:
        moderationStatus:
          type: string
          enum: [active, pending_review, flagged, suspended, removed]
        reason: { type: string, maxLength: 500 }
      required: [moderationStatus]
    Policy:
      type: object
      properties:
        policyType: { type: string }
        title: { type: string }
        content: { type: string }
        version: { type: number }
        publishedAt: { type: ["string", "null"] }
      required: [policyType, title, content, version, publishedAt]
    AdminPolicy:
      allOf:
        - $ref: "#/components/schemas/Policy"
        - type: object
          properties:
            status: { type: string, enum: [draft, published] }
            updatedById: { type: ["string", "null"] }
            updatedAt: { type: string }
          required: [status, updatedById, updatedAt]
    PolicyUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        content: { type: string, minLength: 1, maxLength: 200000 }
        status:
          type: string
          enum: [draft, published]
      required: [title, content]
    SiteSettings:
      type: object
      properties:
        businessName: { type: ["string", "null"] }
        businessAddress: { type: ["string", "null"] }
        supportEmail: { type: ["string", "null"] }
        grievanceOfficerName: { type: ["string", "null"] }
        grievanceOfficerEmail: { type: ["string", "null"] }
        invoiceAdminEmail: { type: ["string", "null"] }
        # A configured tax rate is what turns on the tax breakdown on newly
        # issued invoices -- see taxRatePercent on SiteSettingsUpdate. Null
        # means "no tax configured", not "0% tax".
        taxRatePercent: { type: ["number", "null"] }
        taxRegistrationNumber: { type: ["string", "null"] }
        # Defaults to "Tax" when null -- lets a business call this "VAT",
        # "GST", etc. to match what their region's invoices must say.
        taxLabel: { type: ["string", "null"] }
        # Null means "not configured" -- the Featured/Sponsored checkout
        # falls back to a built-in default price/duration, it is never
        # blocked on an admin having visited Site Settings first.
        sponsorshipPriceCents: { type: ["number", "null"] }
        sponsorshipDurationDays: { type: ["number", "null"] }
      required: [businessName, businessAddress, supportEmail, grievanceOfficerName, grievanceOfficerEmail, invoiceAdminEmail, taxRatePercent, taxRegistrationNumber, taxLabel, sponsorshipPriceCents, sponsorshipDurationDays]
    SiteSettingsUpdate:
      type: object
      properties:
        businessName: { type: ["string", "null"], maxLength: 200 }
        businessAddress: { type: ["string", "null"], maxLength: 500 }
        supportEmail: { type: ["string", "null"], maxLength: 200 }
        grievanceOfficerName: { type: ["string", "null"], maxLength: 200 }
        grievanceOfficerEmail: { type: ["string", "null"], maxLength: 200 }
        invoiceAdminEmail: { type: ["string", "null"], maxLength: 200 }
        taxRatePercent: { type: ["number", "null"], minimum: 0, maximum: 100, multipleOf: 0.01 }
        taxRegistrationNumber: { type: ["string", "null"], maxLength: 100 }
        taxLabel: { type: ["string", "null"], maxLength: 40 }
        sponsorshipPriceCents: { type: ["number", "null"], minimum: 100, maximum: 10000000, multipleOf: 1 }
        sponsorshipDurationDays: { type: ["number", "null"], minimum: 1, maximum: 90, multipleOf: 1 }
    PolicyAcceptanceInput:
      type: object
      properties:
        policyType:
          type: string
          enum: [terms, privacy, website_policy, refund_policy, community_guidelines]
        context:
          type: string
          enum: [listing_submission, campaign_claim, payment, signup]
      required: [policyType, context]
    PolicyAcceptance:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        policyType: { type: string }
        policyVersion: { type: number }
        context: { type: string }
        createdAt: { type: string }
      required: [id, userId, policyType, policyVersion, context, createdAt]
...
    Listing:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        tagline: { type: string }
        description: { type: string }
        category: { type: string }
        initials: { type: string }
        accent: { type: string }
        rank: { type: number }
        effectiveBid: { type: number }
        ownerBid: { type: number }
        communitySupport: { type: number }
        penalties: { type: number }
        supporters: { type: number }
        penalizers: { type: number }
        rating: { type: number }
        reviewCount: { type: number }
        clicks: { type: number }
        clickAnalytics: { $ref: "#/components/schemas/ListingClickAnalytics" }
        createdAt: { type: string }
        trend: { type: number }
        websiteUrl: { type: ["string", "null"] }
        demoVideoUrl:
          type: ["string", "null"]
          description: Optional demo video URL (YouTube, YouTube Shorts, Vimeo, or another http(s) link).
        demoViewCount:
          type: number
          minimum: 0
          description: Count of distinct visitors who have opened this listing's demo video. Always 0 when demoVideoUrl is null.
        logoUrl:
          type: ["string", "null"]
          description: Owner-uploaded brand logo, served from object storage (e.g. `/objects/uploads/<uuid>`). When null, clients should fall back to the initials/accent avatar.
      required:
        - id
        - name
        - slug
        - tagline
        - description
        - category
        - initials
        - accent
        - rank
        - effectiveBid
        - ownerBid
        - communitySupport
        - penalties
        - supporters
        - penalizers
        - rating
        - reviewCount
        - clicks
        - clickAnalytics
        - createdAt
        - trend
        - websiteUrl
        - demoVideoUrl
        - demoViewCount
        - logoUrl
    ListingClickAnalytics:
      type: object
      description: Website click totals for a listing. uniqueClicks counts distinct anonymous visitors; totalClicks counts every recorded click event.
      properties:
        totalClicks: { type: number }
        uniqueClicks: { type: number }
      required:
        - totalClicks
        - uniqueClicks
...
    AdminListing:
      allOf:
        - $ref: "#/components/schemas/Listing"
        - type: object
          properties:
            ownerId: { type: ["string", "null"] }
            status: { type: string, enum: [active, suspended] }
            moderationStatus: { type: string, enum: [active, pending_review, flagged, suspended, removed] }
            moderationReason: { type: ["string", "null"] }
            moderatedAt: { type: ["string", "null"] }
...